rag_minimal/requirements_rag.txt
azure-search-documents==11.6.0
requests>=2.32.3,<3.0.0
